docs(cnpj.json): update CNPJ schema example
I was writing some tests for an application that uses the BrasilAPI CNPJ endpoint following the example shown in the documentation and I kept getting assertion errors. When I made a request to the endpoint, I understood that the documentation is not faithful to its return, so I decided to update the CNPJ schema example so that no one else has to go through this small inconvenience.